requirements.txt
disnake~=2.9.2
h11~=0.14.0
pip~=22.2.1
attrs~=24.2.0
distro~=1.9.0
colorama~=0.4.6
idna~=3.10
multidict~=6.1.0
propcache~=0.2.0
typing_extensions~=4.12.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
httpcore~=1.0.6
certifi~=2024.8.30
jiter~=0.7.0
openai~=1.54.3
httpx~=0.27.2
pydantic~=2.9.2
anyio~=4.6.2.post1
pydantic_core~=2.23.4
yarl~=1.17.1
async-timeout~=4.0.3
aiohttp~=3.10.10
aiosignal~=1.3.1
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.3
annotated-types~=0.7.0
setuptools~=63.2.0
python-dotenv~=1.0.1